Move marathon-specific autoscaling code out of long_running_service_tools into marathon_tools.
